[FE/#60] Refactor : InitialState 모두 Context에서 Reducer로 이동
- Context가 필요 없는 내부 Reducer도 있기 때문에, 상태관리에 관한 초기값은 모두 Reducer에 모아서 관리하도록 변경
- 상태관리 로직이 있는 곳에 초기값도 존재하는게 로직을 알아보기 더 쉬운 거 같다고 판단함
