+++ FE/src/reducers/userReducer.js
@@ -1,3 +1,9 @@
+export const initialState = {
+  id: "sungik-choi",
+  avatarUrl:
+    "https://lh3.googleusercontent.com/lv3q0LxixvhfsBwhLKzpp53EPMQO92jq98uOKqT9sL0hv4gF8l9bl6SRN23FLxkeJH3ldxZt-PreP1zddg",
+};
+
 export const FETCH_SUCCESS = "FETCH_SUCCESS";
 export const FETCH_ERROR = "FETCH_ERROR";
